service-workers/service-worker/websocket-traffic-no-fetch.https.html

<!DOCTYPE html>
<title>Service Worker: WebSocket handshake channel is not intercepted</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js?pipe=sub"></script>
<script src="resources/service-worker-recorder.js"></script>
<script>

async_test(function(t) {
    var path = new URL(".", window.location).pathname
    var url = 'resources/fetch-recording-worker.js';
    var scope = 'resources/blank.html?websocket';
    var host_info = get_host_info();
    var frameURL = host_info['HTTPS_ORIGIN'] + path + scope;
    var frame;

    service_worker_unregister_and_register(t, url, scope)
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated')
            .then(function() {
                return ServiceWorkerRecorder.client.clear(registration.active);
              });
        })
      .then(function() { return with_iframe(frameURL); })
      .then(function(f) {
          frame = f;
          return websocket(t, frame);
        })
      .then(function() {
          return ServiceWorkerRecorder.client.read(
              frame.contentWindow.navigator.serviceWorker.controller
            );
        })
      .then(function(data) {
          for (var url in data) {
            assert_equals(url.indexOf(get_websocket_url()), -1,
                          "Observed an unexpected FetchEvent for the WebSocket handshake");
          }
          t.done();
        })
      .then(function() {
          frame.remove();
          return service_worker_unregister_and_done(t, scope);
        })
      .catch(unreached_rejection(t));
  }, 'Verify WebSocket handshake channel does not get intercepted');
</script>
